1. auto enable scalikejdbc.async.PlayModule in reference.conf 2. move code from PlayModule to PlayInitializer 3. update flyway-play 2.1.0-SNAPSHOT from https://github.com/giabao/flyway-play/tree/auto-enable-PlayModule
