+++ play-plugin/src/main/resources/reference.conf
@@ -1,0 +1,1 @@
+play.modules.enabled += "scalikejdbc.async.PlayModule"
