+++ play-sample/conf/application.conf
@@ -50,18 +50,3 @@ db.default {
   maxIdleMillis=1000
   maxQueueSize=99
 }
-
-# Logger
-# ~~~~~
-# You can also configure logback (http://logback.qos.ch/), by providing a logger.xml file in the conf directory .
-logger {
-  # Root logger:
-  root=DEBUG
-  # Logger used by the framework:
-  play=INFO
-  # Logger provided to your application:
-  application=DEBUG
-}
-
-play.modules.enabled += "org.flywaydb.play.PlayModule"
-play.modules.enabled += "scalikejdbc.async.PlayModule"
